Dashboad loads from db and opens overview

+++ app/src/main/java/com/example/orbitalflightpaths/FlightCalculator.java
@@ -1,54 +1,54 @@
 package com.example.orbitalflightpaths;
 
 class FlightCalculator {
-    private static double uPrimary = 1.32715e+20;
-    private static double parkingOrbitAltitude = 300000;
+    private static float uPrimary = 1.32715e+20f;
+    private static float parkingOrbitAltitude = 300000;
 
 
     FlightCalculator(){
 
     }
 
-    protected double SemiMajorAxis(double orbitradiusS, double orbitRadiusD){
+    protected float SemiMajorAxis(float orbitradiusS, float orbitRadiusD){
         return (orbitradiusS + orbitRadiusD) / 2;
     }
 
-    protected double OrbitVelocity(double orbitRadius){
-        return Math.sqrt(uPrimary * orbitRadius);
+    protected float OrbitVelocity(float orbitRadius){
+        return (float)Math.sqrt(uPrimary * orbitRadius);
     }
 
-    protected double Velocity(double orbitRadius, double semiMajorAxis){
-        return Math.sqrt(uPrimary * ((2 / orbitRadius) - (1 / semiMajorAxis)));
+    protected float Velocity(float orbitRadius, float semiMajorAxis){
+        return (float)Math.sqrt(uPrimary * ((2 / orbitRadius) - (1 / semiMajorAxis)));
     }
 
-    protected double VelocityInf(double velocity, double orbitVelocity){
+    protected float VelocityInf(float velocity, float orbitVelocity){
         return Math.abs(velocity - orbitVelocity);
     }
 
-    protected double uS(double planetMass){
-        return (6.674e-11) * planetMass;
+    protected float uS(float planetMass){
+        return (6.674e-11f) * planetMass;
     }
 
-    protected double ParkingOrbitRadius(double planetRadius){
+    protected float ParkingOrbitRadius(float planetRadius){
         return planetRadius + parkingOrbitAltitude;
     }
 
-    protected double ParkingOrbitCircularVel(double uS, double parkingOrbitRadius){
-        return Math.sqrt(uS/parkingOrbitRadius);
+    protected float ParkingOrbitCircularVel(float uS, float parkingOrbitRadius){
+        return (float)Math.sqrt(uS/parkingOrbitRadius);
     }
 
-    protected double VelocityEscape(double uS, double parkingOrbitRadius){
-        return Math.sqrt((2 * uS)/parkingOrbitRadius);
+    protected float VelocityEscape(float uS, float parkingOrbitRadius){
+        return (float)Math.sqrt((2 * uS)/parkingOrbitRadius);
     }
 
-    protected double VelocityHyper(double velocityInf, double velocityEsc){
-        return Math.sqrt((velocityInf *velocityInf) + (velocityEsc *velocityEsc));
+    protected float VelocityHyper(float velocityInf, float velocityEsc){
+        return (float)Math.sqrt((velocityInf *velocityInf) + (velocityEsc *velocityEsc));
     }
 
-    protected double DeltaV(double velocityHyper, double parkingOrbitCircularVel){
+    protected float DeltaV(float velocityHyper, float parkingOrbitCircularVel){
         return velocityHyper - parkingOrbitCircularVel;
     }
-    protected double FinalDeltaV(double deltavS, double deltavD){
+    protected float FinalDeltaV(float deltavS, float deltavD){
         return (Math.abs(deltavS) + Math.abs(deltavD));
     }
 }
